[#25] style: indicator 추가
데이터를 받아와서 tableview를 리로드하면서 indicator 사라지도록 함.
추가로 경도, 위도를 double형으로 보내도록 변경

+++ iOS/Dust/Dust/Model/DataTask.swift
@@ -12,7 +12,7 @@ class DataTask {
     let defaultSession = URLSession(configuration: .default)
     var dataTask: URLSessionDataTask?
     
-    func requestInfoFromNearStation(latitude: Int, longitude: Int) {
+    func requestInfoFromNearStation(latitude: Double, longitude: Double) {
         guard var urlComponent = URLComponents(string: "http://13.124.46.74:8080/dust-status") else { return }
         let latitudeQuery = URLQueryItem(name: "latitude", value: String(latitude))
         let longtitudeQuery = URLQueryItem(name: "longitude", value: String(longitude))
